Migrate languages (#139)
* Migrate language manifests to make them importable (#137)

The language manifests are migrated into a different location, such that
they will be imported automatically. The json content has been regenerated
by use of the developer console, which is introducing a version number.

Minor adjustments have been made to names:

skill.json: "Alexa actionable notifications"
invocation names: "actionable notifications"

Both are not used to work with this skill. The skill is triggered from the
end of Home Assistant by using the **skill id**.

"actionable notifications" is set for all languages. This may be translated
in future, albeit not really needed, as pointed out just before.

Note 1:

The importer always shows a warning, that the invocation name must not
be identical with the skill name. It does occure no matter what combinations
of names you give. I tested for a whole hour. It's oviously a bug of the
importer.

It always does reset the invocation name to "change me". You actually have
to set it to something else again. The TEST simulater in the developer console
will not detect the skill, if you call "open change me". Use anything
else i.e. 'actionable notifications'.

Note 2:

I noticed that it is not enough to provide a language manifest file. There
has to be an entry in skill.json as well. In this case it is possible to
directly import this language without having to import en-US first.

Note 3:

Most manifests are not really translated and still hold english
expressions. Doing translations is out of the scope of this commit.

* Updated languages

Co-authored-by: Elmar Hinz <elmar-hinz@users.noreply.github.com>